portfolio/requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
alpha-vantage==2.3.1
altair==5.3.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
cassandra-driver==3.29.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
DateTime==5.5
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.20.0
feedparser==6.0.11
flatbuffers==24.3.25
fonttools==4.52.4
fqdn==1.5.1
frozendict==2.4.4
frozenlist==1.4.1
gast==0.4.0
geomet==0.2.1.post1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.18.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keras==2.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
multidict==6.0.5
multitasking==0.0.11
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandas-datareader==0.10.0
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
patsy==0.5.6
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
pyspark==3.5.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytrends==4.9.2
pytz==2024.1
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
stocknews==0.9.11
streamlit==1.37.0
streamlit-tags==1.2.8
tenacity==8.3.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
termcolor==2.4.0
terminado==0.18.1
textblob==0.18.0.post0
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.5.0
tzdata==2024.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.1
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wordcloud==1.9.3
wrapt==1.16.0
yarl==1.9.4
yfinance==0.2.40
zipp==3.19.0
zope.interface==6.4.post2
